Revert "Int64 updated to Float64 in schemas (#82)" (#85)
This reverts commit d5189cdfe443cdc7ad86acbf51e47af732eeb7c1.

+++ model.go
@@ -163,7 +163,7 @@ const FlogJsonSchema string = `{
     "fields": [
         {
             "name": "bytes",
-            "data_type": "Float64",
+            "data_type": "Int64",
             "nullable": true,
             "dict_id": 0,
             "dict_is_ordered": false,
@@ -248,7 +248,7 @@ const FlogJsonSchema string = `{
         },
         {
             "name": "status",
-            "data_type": "Float64",
+            "data_type": "Int64",
             "nullable": true,
             "dict_id": 0,
             "dict_is_ordered": false,
@@ -278,7 +278,7 @@ const SchemaBody string = `{
         },
         {
             "name": "device_id",
-            "data_type": "Float64",
+            "data_type": "Int64",
             "nullable": true,
             "dict_id": 0,
             "dict_is_ordered": false,
@@ -355,7 +355,7 @@ const SchemaBody string = `{
         },
         {
             "name": "process_id",
-            "data_type": "Float64",
+            "data_type": "Int64",
             "nullable": true,
             "dict_id": 0,
             "dict_is_ordered": false,
@@ -371,7 +371,7 @@ const SchemaBody string = `{
         },
         {
             "name": "response_time",
-            "data_type": "Float64",
+            "data_type": "Int64",
             "nullable": true,
             "dict_id": 0,
             "dict_is_ordered": false,
@@ -408,7 +408,7 @@ const SchemaBody string = `{
         },
         {
             "name": "status_code",
-            "data_type": "Float64",
+            "data_type": "Int64",
             "nullable": true,
             "dict_id": 0,
             "dict_is_ordered": false,
@@ -432,7 +432,7 @@ const SchemaBody string = `{
         },
         {
             "name": "user_id",
-            "data_type": "Float64",
+            "data_type": "Int64",
             "nullable": true,
             "dict_id": 0,
             "dict_is_ordered": false,
